Add mediaservice
To support operations like DELETE or COPY, we need to add a service layer that performs
filesystem operations alongside database operations.

This commit adds a fileoperator package, localmediafile operator,
a media service and a new /DELETE endpoint to the media item handler.

Add recyclebin/* to gitignore for deleted item store